LayoutTests/imported/w3c/web-platform-tests/svg/geometry/svg-get-bounding-client-rect-in-non-rendered-elements-expected.txt


FAIL Get Bounding Client Rect assert_equals: rect1 expected 0 but got 10
